+++ test/eachDeep.test.js
@@ -8,44 +8,56 @@ const chai = require('chai'),
 const asserttype = require('chai-asserttype');
 chai.use(asserttype);
 var { demo, circular } = require('./object');
+var { validateIteration } = require('./common.js');
 
 describe('eachDeep', () => {
   it('no mutation', () => {
     let orig = _.cloneDeep(demo);
     let obj = _.cloneDeep(demo);
-    _.eachDeep(obj, (value, key) => {
+    _.eachDeep(obj, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
       if (key == 'skip') return false;
     });
     expect(obj).to.deep.equal(orig);
   });
-  it('Count nodes', () => {
-    let c = 0;
-    _.eachDeep(demo, (value, key) => {
+  it('eachDeep defaults', () => {
+    let keys = [];
+    _.eachDeep(demo, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      keys.push(key);
       if (key == 'skip') return false;
-      c++;
     });
-    expect(c).equal(25);
+
+    expect(JSON.stringify(keys)).equal(
+      '[null,"a","b","c","d","0","i","1","i","2","i","3","i","4","i","5","i","6","o","d","f","skip","s","b","n","u","nl"]'
+    );
   });
   it('Chaining', () => {
-    let c = 0;
+    let keys = [];
     _(demo)
-      .eachDeep((value, key) => {
+      .eachDeep((value, key, parent, ctx) => {
+        validateIteration(value, key, parent, ctx);
+        keys.push(key);
         if (key == 'skip') return false;
-        c++;
       })
       .value();
-    expect(c).equal(25);
+    expect(JSON.stringify(keys)).equal(
+      '[null,"a","b","c","d","0","i","1","i","2","i","3","i","4","i","5","i","6","o","d","f","skip","s","b","n","u","nl"]'
+    );
   });
   it('alias forEachDeep', () => {
-    let c = 0;
-    _.forEachDeep(demo, () => {
-      c++;
+    let keys = [];
+    _.forEachDeep(demo, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      keys.push(key);
     });
-    expect(c).equal(30);
+    expect(JSON.stringify(keys)).equal(
+      '[null,"a","b","c","d","0","i","1","i","2","i","3","i","4","i","5","i","6","o","d","f","skip","please","dont","go","here","s","b","n","u","nl"]'
+    );
   });
   it('returns collection', () => {
-    let c = _.forEachDeep(demo, () => {
-      /*hello?*/
+    let c = _.forEachDeep(demo, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
     });
     expect(c).equal(demo);
   });
@@ -60,102 +72,116 @@ describe('eachDeep', () => {
 
   it('circular', () => {
     let circluarPath = [];
+    let options = { checkCircular: true };
     _.eachDeep(
       circular,
       (value, key, parent, ctx) => {
+        validateIteration(value, key, parent, ctx, options);
         if (ctx.isCircular) {
           circluarPath.push(ctx.path);
 
           return false;
         }
       },
-      { checkCircular: true }
+      options
     );
-
-    expect(circluarPath)
-      .to.include('a.b.c.e')
-      .and.to.include('i[5][0]')
-      .and.to.include('i[5][1][0].b.c.e')
-      .and.to.have.property('length')
-      .and.equal(3);
+    expect(JSON.stringify(circluarPath)).equal(
+      '["a.b.c.e","i[5][0]","i[5][1][0].b.c.e"]'
+    );
   });
 
   it('circular with array path format', () => {
     let circluarPath = [];
+    let options = { checkCircular: true, pathFormat: 'array' };
     _.eachDeep(
       circular,
       (value, key, parent, ctx) => {
-        ctx.parents.forEach((p) => {
-          expect(p.path).to.be.an.array();
+        validateIteration(value, key, parent, ctx, options);
+        ctx.parents.forEach((p, i) => {
+          if (i) {
+            expect(p.path).to.be.an.array();
+          } else {
+            expect(p.path).equal(undefined);
+          }
         });
-        if (_.findIndex(ctx.parents, ['value', value]) != -1) {
+        if (ctx.isCircular) {
           circluarPath.push(_.pathToString(ctx.path));
-
           return false;
         }
       },
-      { pathFormat: 'array' }
+      options
     );
-
-    expect(circluarPath)
-      .to.include('a.b.c.e')
-      .and.to.include('i[5][0]')
-      .and.to.include('i[5][1][0].b.c.e')
-      .and.to.have.property('length')
-      .and.equal(3);
+    expect(JSON.stringify(circluarPath)).equal(
+      '["a.b.c.e","i[5][0]","i[5][1][0].b.c.e"]'
+    );
   });
 
   it('Array', () => {
-    let c = 0;
-    _.eachDeep([demo, demo], (value, key) => {
+    let keys = [];
+    _.eachDeep([demo, demo], (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      keys.push(key);
       if (key == 'skip') return false;
-      c++;
     });
-    expect(c).equal(52);
+    expect(JSON.stringify(keys)).equal(
+      '["0","a","b","c","d","0","i","1","i","2","i","3","i","4","i","5","i","6","o","d","f","skip","s","b","n","u","nl","1","a","b","c","d","0","i","1","i","2","i","3","i","4","i","5","i","6","o","d","f","skip","s","b","n","u","nl"]'
+    );
   });
 
   it('String', () => {
-    let c = 0;
-    _.eachDeep('Hello?', (value, key) => {
-      if (key == 'skip') return false;
-      c++;
+    let paths = [];
+    _.eachDeep('Hello?', (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      paths.push(ctx.path);
     });
-    expect(c).equal(0);
+    expect(paths).deep.equal([undefined]);
   });
+
   it('empty props', () => {
-    let c = 0;
+    let paths = [];
     var o = { a: 0, b: 1, c: 2 };
     delete o.b;
-    _.eachDeep(o, () => {
-      c++;
+    _.eachDeep(o, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      paths.push(ctx.path);
     });
-    expect(c).equal(2);
-    c = 0;
+    expect(paths).deep.equal([undefined, 'a', 'c']);
+    paths = [];
     var a = ['a', 'b', 'c'];
     delete a[1];
-    _.eachDeep(a, () => {
-      c++;
+    _.eachDeep(a, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      paths.push(ctx.path);
     });
-    expect(c).equal(2);
-    c = 0;
+    expect(paths).deep.equal(['[0]', '[2]']);
+    paths = [];
     var slots = ['start', , 'middle', , 'finish'];
-    _.eachDeep(slots, () => {
-      c++;
+    _.eachDeep(slots, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      paths.push(ctx.path);
     });
-    expect(c).equal(3);
+    expect(paths).deep.equal(['[0]', '[2]', '[4]']);
   });
   it('generated string paths are correct', () => {
     _.eachDeep(demo, function(value, key, parent, ctx) {
-      assert(_.has(demo, ctx.path), 'Incorrect path: ' + ctx.path);
+      validateIteration(value, key, parent, ctx);
+      if (parent) {
+        assert(_.exists(demo, ctx.path), 'Incorrect path: ' + ctx.path);
+      }
     });
   });
   it('generated array paths are correct', () => {
+    let options = { pathFormat: 'array' };
     _.eachDeep(
       demo,
       function(value, key, parent, ctx) {
-        assert(_.has(demo, ctx.path), 'Incorrect path: ' + ctx.path);
+        validateIteration(value, key, parent, ctx, options);
+        assert(
+          _.obtain(demo, ctx.path) === value,
+          'Incorrect path: ' + ctx.path
+        );
       },
-      { pathFormat: 'array' }
+      options
     );
   });
   it('empty slots skipped', () => {
@@ -166,8 +192,9 @@ describe('eachDeep', () => {
       [, , , 'a'],
       {},
     ];
-    _.eachDeep(obj, (v) => {
-      expect(v).not.to.equal(undefined);
+    _.eachDeep(obj, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      expect(value).not.to.equal(undefined);
     });
   });
   it('undefined are valid values, no skip', () => {
@@ -185,8 +212,9 @@ describe('eachDeep', () => {
       {},
     ];
     let u = 0;
-    _.eachDeep(obj, (v) => {
-      if (v === undefined) {
+    _.eachDeep(obj, (value, key, parent, ctx) => {
+      validateIteration(value, key, parent, ctx);
+      if (value === undefined) {
         u++;
       }
     });