+++ test/filterDeep.test.js
@@ -8,93 +8,365 @@ const chai = require('chai'),
 
 const asserttype = require('chai-asserttype');
 chai.use(asserttype);
+var { validateIteration } = require('./common.js');
 
 var { demo, circular } = require('./object');
-function isNS(value) {
-  let t = typeof value;
-  return t == 'number' || t == 'string';
+function isNS(options = {}) {
+  return (value, key, parent, ctx) => {
+    // console.log(`@${ctx.path}`);
+    options.method = 'filterDeep';
+    validateIteration(value, key, parent, ctx, options);
+    let t = typeof value;
+    return t == 'number' || t == 'string';
+  };
 }
 
 describe('filterDeep', () => {
   it('no mutation', () => {
     let orig = _.cloneDeep(demo);
     let obj = _.cloneDeep(demo);
-    _.filterDeep(obj, isNS, { leavesOnly: true });
+    _.filterDeep(obj, isNS());
     expect(obj).to.deep.equal(orig);
   });
-  it('Count paths', () => {
+  it('filter object - defaults', () => {
     let filtrate = _(demo)
-      .filterDeep(isNS, { leavesOnly: true })
-      .index({ leavesOnly: true })
+      .filterDeep(isNS())
       .value();
-    // console.log(_.filterDeep(demo, isNS));
-    expect(_.size(filtrate)).equal(9);
+    expect(JSON.stringify(filtrate)).equal(
+      '{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{"skip":{"please":{"dont":{"go":{"here":"skip it"}}}}}}],"s":"hello"}},"n":12345}}'
+    );
   });
-  it('Array', () => {
-    let filtrate = _.filterDeep([demo, demo], isNS, { leavesOnly: true });
+  it('filter object - no clone', () => {
+    let options = { cloneDeep: false };
+    let filtrate = _(demo)
+      .filterDeep(isNS(options), options)
+      .value();
+    expect(JSON.stringify(filtrate)).equal(
+      '{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{"skip":{"please":{"dont":{"go":{"here":"skip it"}}}}}}],"s":"hello"}},"n":12345}}'
+    );
+  });
+
+  it('filter array - no clone', () => {
+    let options = { cloneDeep: false };
+    let filtrate = _([demo])
+      .filterDeep(isNS(options), options)
+      .value();
+    expect(JSON.stringify(filtrate)).equal(
+      '[{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{"skip":{"please":{"dont":{"go":{"here":"skip it"}}}}}}],"s":"hello"}},"n":12345}}]'
+    );
+  });
+
+  it('filter object - no clone, not leavesOnly', () => {
+    let options = {
+      cloneDeep: false,
+      leavesOnly: false,
+      onFalse: { skipChildren: false },
+    };
+    let filtrate = _(demo)
+      .filterDeep(isNS(options), options)
+      .value();
+    expect(JSON.stringify(filtrate)).equal(
+      '{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{"skip":{"please":{"dont":{"go":{"here":"skip it"}}}}}}],"s":"hello"}},"n":12345}}'
+    );
+  });
+
+  it('filter array - no clone, not leavesOnly', () => {
+    let options = {
+      cloneDeep: false,
+      leavesOnly: false,
+      onFalse: { skipChildren: false },
+    };
+    let filtrate = _([demo])
+      .filterDeep(isNS(options), options)
+      .value();
+    expect(JSON.stringify(filtrate)).equal(
+      '[{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{"skip":{"please":{"dont":{"go":{"here":"skip it"}}}}}}],"s":"hello"}},"n":12345}}]'
+    );
+  });
+
+  it('filter array - no clone, not leavesOnly, includeRoot', () => {
+    let options = {
+      cloneDeep: false,
+      leavesOnly: false,
+      onFalse: { skipChildren: false },
+      includeRoot: true,
+    };
+    let filtrate = _([demo])
+      .filterDeep(isNS(options), options)
+      .value();
+    expect(JSON.stringify(filtrate)).equal(
+      '[{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{"skip":{"please":{"dont":{"go":{"here":"skip it"}}}}}}],"s":"hello"}},"n":12345}}]'
+    );
+  });
+
+  it('filter object -  array', () => {
+    let filtrate = _.filterDeep([demo, demo], isNS());
     expect(filtrate).to.be.an('array');
-    filtrate = _.index(filtrate, { leavesOnly: true });
-    expect(_.size(filtrate)).equal(18);
+    expect(JSON.stringify(filtrate)).equal(
+      '[{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{"skip":{"please":{"dont":{"go":{"here":"skip it"}}}}}}],"s":"hello"}},"n":12345}},{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{"skip":{"please":{"dont":{"go":{"here":"skip it"}}}}}}],"s":"hello"}},"n":12345}}]'
+    );
   });
   it('Completely filtered out', () => {
     let obj = { a: { b: undefined } };
-    let filtrate = _.filterDeep(obj, isNS, {
-      leavesOnly: true,
-    });
+    let filtrate = _.filterDeep(obj, isNS());
     //console.log(filtrate);
-    expect(filtrate).to.deep.equal({});
+    expect(filtrate).to.deep.equal(null);
+    obj = [{ a: { b: undefined } }];
+    filtrate = _.filterDeep(obj, isNS());
+    //console.log(filtrate);
+    expect(filtrate).to.deep.equal(null);
   });
-  it('Count paths circular', () => {
-    let filtrate = _.filterDeep(circular, isNS, {
-      checkCircular: true,
-      leavesOnly: true,
-    });
-    // console.log(circular);
-    // console.log(filtrate);
-    filtrate = _.paths(filtrate, {
-      leavesOnly: true,
-      checkCircular: true,
-    });
-    // console.log(filtrate);
-    expect(_.size(filtrate)).equal(11);
+
+  it('I dunno', () => {
+    let filtrate = _.filterDeep(demo, (value, key, parent, ctx) =>
+      validateIteration(value, key, parent, ctx)
+    );
+    //console.log(filtrate);
+    expect(filtrate).equal(null);
+    filtrate = _.filterDeep([demo], (value, key, parent, ctx) =>
+      validateIteration(value, key, parent, ctx)
+    );
+    //console.log(filtrate);
+    expect(filtrate).equal(null);
   });
+
   it('Chaining', () => {
+    let options = {
+      leafsOnly: false,
+      onTrue: { skipChildren: false },
+    };
     let filtrate = _(demo)
-      .filterDeep(isNS, { leafsOnly: true })
-      .paths({ leavesOnly: true })
+      .filterDeep(isNS())
+      .filterDeep((value, key, parent, ctx) => {
+        validateIteration(value, key, parent, ctx, options);
+        return key !== 'skip';
+      }, options)
       .value();
-    expect(_.size(filtrate)).equal(9);
+    expect(JSON.stringify(filtrate)).equal(
+      '{"a":{"b":{"c":{"d":[{"i":0},{"i":1},{"i":2},{"i":3},{"i":4},{"i":5},{"o":{}}],"s":"hello"}},"n":12345}}'
+    );
   });
 
-  it('returns empty obj', () => {
-    let filtrate = _.filterDeep(1, () => true);
-    expect(filtrate).to.deep.equal({});
+  it('non-object', () => {
+    expect(
+      _.filterDeep(
+        true,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(true);
+    expect(
+      _.filterDeep(
+        true,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
 
-    filtrate = _.filterDeep('123', () => true);
-    expect(filtrate).to.deep.equal({});
+    expect(
+      _.filterDeep(
+        false,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(false);
+    expect(
+      _.filterDeep(
+        false,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
 
-    filtrate = _.filterDeep(null, () => true);
-    expect(filtrate).to.deep.equal({});
+    expect(
+      _.filterDeep(
+        null,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(null);
+    expect(
+      _.filterDeep(
+        null,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
 
-    filtrate = _.filterDeep(undefined, () => true);
-    expect(filtrate).to.deep.equal({});
+    expect(
+      _.filterDeep(
+        undefined,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(undefined);
+    expect(
+      _.filterDeep(
+        undefined,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
 
-    filtrate = _.filterDeep(() => {}, () => true);
-    expect(filtrate).to.deep.equal({});
+    expect(
+      _.filterDeep(
+        1,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(1);
+    expect(
+      _.filterDeep(
+        1,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
 
-    filtrate = _.filterDeep(new Date(), () => true);
-    expect(filtrate).to.deep.equal({});
+    expect(
+      _.filterDeep(
+        0,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(0);
+    expect(
+      _.filterDeep(
+        0,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
 
-    filtrate = _.filterDeep(/.*/, () => true);
-    expect(filtrate).to.deep.equal({});
+    expect(
+      _.filterDeep(
+        'hi',
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal('hi');
+    expect(
+      _.filterDeep(
+        'hi',
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
+
+    expect(
+      _.filterDeep(
+        {},
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal({});
+    expect(
+      _.filterDeep(
+        {},
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
+
+    let options = { includeRoot: true };
+    expect(
+      _.filterDeep(
+        [],
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx, options) || true,
+        options
+      )
+    ).to.deep.equal([]);
+
+    expect(
+      _.filterDeep(
+        [],
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(null);
+    expect(
+      _.filterDeep(
+        [],
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
+
+    expect(
+      _.filterDeep(
+        () => {},
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal({});
+    expect(
+      _.filterDeep(
+        () => {},
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
+
+    let dt = new Date();
+    expect(
+      _.filterDeep(
+        dt,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(dt);
+    expect(
+      _.filterDeep(
+        dt,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
+
+    let rx = /.*/;
+    expect(
+      _.filterDeep(
+        rx,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(rx);
+    expect(
+      _.filterDeep(
+        rx,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
+
+    let sm = Symbol('Halloo');
+    expect(
+      _.filterDeep(
+        sm,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || true
+      )
+    ).to.deep.equal(sm);
+    expect(
+      _.filterDeep(
+        sm,
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) && false
+      )
+    ).to.deep.equal(null);
   });
 
   it('Obj only', () => {
     let filtrate = _(demo)
-      .filterDeep(_.isObject, {
-        leavesOnly: false,
-        onTrue: { skipChildren: false },
-      })
+      .filterDeep(
+        (value, key, parent, ctx) =>
+          validateIteration(value, key, parent, ctx) || _.isObject(value),
+        {
+          leavesOnly: false,
+          onTrue: { skipChildren: false },
+        }
+      )
       .omitDeep('o.d', { onMatch: {} })
       .value();
     // console.log(JSON.stringify(filtrate));
@@ -122,26 +394,42 @@ describe('filterDeep', () => {
     });
   });
   it('Circular', () => {
-    let filtrate = _.filterDeep(circular, isNS, {
+    let options = {
       checkCircular: true,
       leavesOnly: true,
-    });
+    };
+    let filtrate = _.filterDeep(circular, isNS(options), options);
+
     filtrate.should.have.nested.property('a.b.c.e').and.equal(filtrate.a.b);
-    // console.log(_.get(filtrate, 'i[5][1][0].b'));
-    // console.log(_.get(filtrate, 'a.b'));
-    // console.log(circular);
-    //console.log(filtrate);
     filtrate.should.have.nested.property('i[5][0]').and.equal(filtrate);
     filtrate.should.have.nested
       .property('i[5][1][0].b.c.e')
       .and.equal(filtrate.i[5][1][0].b);
+
+    filtrate = _.paths(filtrate, options);
+
+    expect(filtrate).to.deep.equal([
+      'a.b.c.hi',
+      'a.b.c.e',
+      'i[0]',
+      'i[1]',
+      'i[2]',
+      'i[3]',
+      'i[4].hello',
+      'i[5][0]',
+      'i[5][1][0].b.c.hi',
+      'i[5][1][0].b.c.e',
+      'i[5][2].j.hello',
+    ]);
   });
+
   it("Don't keep circular", () => {
-    let filtrate = _.filterDeep(circular, isNS, {
+    let options = {
       checkCircular: true,
       keepCircular: false,
       leavesOnly: true,
-    });
+    };
+    let filtrate = _.filterDeep(circular, isNS(options), options);
     let err;
     try {
       JSON.stringify(filtrate);
@@ -157,10 +445,11 @@ describe('filterDeep', () => {
     obj.x = { y: { z: {} } };
     obj.x.y.z.c = obj.x;
     obj.x.y.z1 = obj.x.y;
-    let filtrate = _.filterDeep(obj, isNS, {
+    let options = {
       checkCircular: true,
       leavesOnly: true,
-    });
+    };
+    let filtrate = _.filterDeep(obj, isNS(options), options);
     // console.log(filtrate);
     filtrate.should.not.have.property('x');
   });
@@ -177,20 +466,22 @@ describe('filterDeep', () => {
   });
   it('replaceCircularBy', () => {
     let again = Symbol('[circular]');
-    let filtrate = _.filterDeep(circular, isNS, {
+    let options = {
       checkCircular: true,
       leavesOnly: true,
       replaceCircularBy: again,
-    });
+    };
+    let filtrate = _.filterDeep(circular, isNS(options), options);
     // console.log(filtrate);
     filtrate.should.have.nested.property('a.b.c.e').and.equal(again);
     filtrate.should.have.nested.property('i[5][0]').and.equal(again);
     filtrate.should.have.nested.property('i[5][1][0].b.c.e').and.equal(again);
-    filtrate = _.filterDeep(circular, isNS, {
+    options = {
       checkCircular: true,
       leavesOnly: true,
       replaceCircularBy: undefined,
-    });
+    };
+    filtrate = _.filterDeep(circular, isNS(options), options);
     // console.log(filtrate);
     filtrate.should.have.nested.property('a.b.c.e').and.equal(undefined);
     filtrate.should.have.nested.property('i[5][0]').and.equal(undefined);